Implemented expansion panel in ViewProfileActivity,  Updated RegisterActivity, launchScreenActivity, OrderSummary...
- Implemented expansion panel in ViewProfileActivity,
- Updated RegisterActivity, launchScreenActivity,
- Change accountPoints and awardedPoints in OrderSummaryRecord, PurchaseAvtivity, FoodItemActivity
- FoodItemActivity handles points accumulation and redeem functionality

+++ app/src/main/java/com/example/steadykopitiam/OrderSummaryRecord.kt
@@ -2,7 +2,7 @@ package com.example.steadykopitiam
 
 class OrderSummaryRecord(
     val orderSummaryTimeDate : String,
-    val orderSummaryAwardedPoints : Double,
+    val orderSummaryAwardedPoints : Int,
     val orderSummarySubtotal : Double,
     val orderSummaryFoodName  : String ,
     val orderSummaryFoodCarbs : Int,
